Fix negative list indexing (lst[-1]) for homogeneous fast-path lists
Negative indexing on a list — lst[-1], one of the most common indexing
idioms in Python — either raised a bogus IndexError or silently
produced/wrote the wrong value, depending on the list's internal
storage representation.

Root cause: Codegen.cpp's A4/A7 optimization routes subscripts on a
compile-time-proven homogeneous int/float list straight to native
fast-path calls (PyList_GetItemInt64/Double, PyList_SetItemInt64/Double,
and their Auto variants) instead of the boxed Pyc_Subscript path. Unlike
every other indexing path in the runtime (str, bytes, the generic
Pyc_Subscript/PyList_GetItemObj/PyList_GetItemI64 path), these six
functions took an unsigned size_t index with no "if negative, add
length" normalization — a negative i64 index reinterpreted as size_t
becomes an enormous value, failing every bounds check. Get functions hit
the out-of-range fallback and raised a bogus IndexError; set functions
don't raise, so they silently no-op'd instead.

Fixed with a shared pyc_normalize_list_index helper applied in all six
functions, and changed their signatures from size_t to a signed long so
a negative index survives the call (the LLVM side already declares the
parameter as a plain i64 with no signedness of its own, so no
Codegen.cpp changes were needed).

Verified against real CPython for get/set on int and float lists,
various negative indices, out-of-range still raising IndexError, and a
list passed through an untyped function parameter. Added as permanent
tests/runner.py regressions (350/350, up from 345), import tests still
green (9/9), valgrind shows 0 errors.

Also documents (not fixed, minor) that a homogeneous bool list loses
its bool-ness on read (prints 1/0 instead of True/False), and refines
the existing **kwargs documentation with a new finding: a nested
function with both *args and **kwargs in its signature crashes
compilation entirely — breaking the standard generic-decorator wrapper
pattern (def wrapper(*args, **kwargs): return f(*args, **kwargs)).

+++ include/pyc/runtime.h
@@ -43,12 +43,12 @@ int PyList_Unpack3(PyObject* list, PyObject** out0, PyObject** out1, PyObject** 
 PyObject* PyList_NewBoxed(PyObject* n);
 PyObject* PyList_NewIntBoxed(PyObject* n);
 PyObject* PyList_NewFloatBoxed(PyObject* n);
-long      PyList_GetItemInt64(PyObject* list, size_t index);
-double    PyList_GetItemDouble(PyObject* list, size_t index);
-void      PyList_SetItemInt64(PyObject* list, size_t index, long v);
-void      PyList_SetItemDouble(PyObject* list, size_t index, double v);
-void      PyList_SetItemInt64Auto(PyObject* list, size_t index, long v);
-void      PyList_SetItemDoubleAuto(PyObject* list, size_t index, double v);
+long      PyList_GetItemInt64(PyObject* list, long index);
+double    PyList_GetItemDouble(PyObject* list, long index);
+void      PyList_SetItemInt64(PyObject* list, long index, long v);
+void      PyList_SetItemDouble(PyObject* list, long index, double v);
+void      PyList_SetItemInt64Auto(PyObject* list, long index, long v);
+void      PyList_SetItemDoubleAuto(PyObject* list, long index, double v);
 void      PyList_SetItemBoxed(PyObject* list, PyObject* idx, PyObject* item);
 PyObject* PyList_Comprehension(int start, int end);
 PyObject* PyFloat_FromDouble(double v);
